feat: refactored agents create into ai and etl crates
