+++ phymes-agents/README.md
@@ -6,6 +6,6 @@ Agents crate
 
 ## Synopsis
 
-The PHYMES agent crate implements the functionality to enable Agentic AI. Specifically, the crate implements functionality for native rust text generation inference, text embedding inference, and general tensor operations using the [candle](https://github.com/huggingface/candle) crates from [HuggingFace](https://huggingface.co/). All functionality can be GPU accelerated using CUDA and CuDNN.
+The PHYMES agent crate implements the functionality to enable Agentic AI on desktop, mobile, and in the browser. Templates to create a chatbot, tool calling agent, and document RAG agent are provided and will be expanded in the future.
 
 <!--- ANCHOR_END: synopsis --->
