--- phymes-agents/src/candle_models/mod.rs
@@ -1,2 +1,0 @@
-pub mod quantized_bert;
-pub mod quantized_qwen2;
