+++ phymes-agents/src/lib.rs
@@ -1,7 +1,1 @@
-pub mod candle_assets;
-pub mod candle_chat;
-pub mod candle_embed;
-pub mod candle_models;
-pub mod candle_ops;
-pub mod openai_asset;
 pub mod session_plans;
