+++ phymes-agents/src/session_plans/document_rag_session.rs
@@ -2,19 +2,18 @@ use anyhow::Result;
 use std::sync::Arc;
 
 #[cfg(feature = "openai_api")]
-use crate::openai_asset::{
+use phymes_ai::openai_asset::{
     chat_processor::OpenAIChatProcessor, embed_processor::OpenAIEmbedProcessor,
 };
-use crate::{
-    candle_chat::{
+use phymes_ai::{
+    candle_assets::candle_which::WhichCandleAsset, candle_chat::{
         chat_config::CandleChatConfig, chat_processor::CandleChatProcessor,
         message_aggregator_processor::MessageAggregatorProcessor,
-    },
-    candle_embed::{embed_config::CandleEmbedConfig, embed_processor::CandleEmbedProcessor},
-    candle_ops::{
-        ops_config::CandleOpsConfig, ops_processor::CandleOpProcessor, ops_which::WhichCandleOps,
-        summary_config::CandleOpsSummaryConfig, summary_processor::OpsSummaryProcessor,
-    },
+    }, candle_embed::{embed_config::CandleEmbedConfig, embed_processor::CandleEmbedProcessor}
+};
+use phymes_etl::candle_ops::{
+    ops_config::CandleOpsConfig, ops_processor::CandleOpProcessor, ops_which::WhichCandleOps,
+    summary_config::CandleOpsSummaryConfig, summary_processor::OpsSummaryProcessor,
 };
 use phymes_core::{
     metrics::ArrowTaskMetricsSet,
@@ -647,7 +646,7 @@ impl AgentSessionBuilderTrait for DocumentRAGSession<'_> {
                 std::env::var("HOME").unwrap_or("".to_string())
             )),
             candle_asset: Some(
-                crate::candle_assets::candle_which::WhichCandleAsset::QwenV2p5_1p5bChat,
+                WhichCandleAsset::QwenV2p5_1p5bChat,
             ),
             ..Default::default()
         };
@@ -656,7 +655,7 @@ impl AgentSessionBuilderTrait for DocumentRAGSession<'_> {
         #[cfg(feature = "hf_hub")]
         {
             candle_chat_config.candle_asset =
-                Some(crate::candle_assets::candle_which::WhichCandleAsset::QwenV2p5_3bChat);
+                Some(WhichCandleAsset::QwenV2p5_3bChat);
             candle_chat_config.openai_asset = None;
             candle_chat_config.weights_config_file = None;
             candle_chat_config.weights_file = None;
@@ -706,8 +705,8 @@ impl AgentSessionBuilderTrait for DocumentRAGSession<'_> {
                 std::env::var("HOME").unwrap_or("".to_string())
             )),
             candle_asset: Some(
-                // crate::candle_assets::candle_which::WhichCandleAsset::BertEmbed,
-                crate::candle_assets::candle_which::WhichCandleAsset::QuantizedBertEmbed,
+                // WhichCandleAsset::BertEmbed,
+                WhichCandleAsset::QuantizedBertEmbed,
             ),
             // weights_config_file: Some(format!(
             //     "{}/.cache/hf/models--Alibaba-NLP--gte-Qwen2-1.5B-instruct/config.json",
@@ -726,7 +725,7 @@ impl AgentSessionBuilderTrait for DocumentRAGSession<'_> {
             //     std::env::var("HOME").unwrap_or("".to_string())
             // )),
             // candle_asset: Some(
-            //     crate::candle_assets::candle_which::WhichCandleAsset::QwenV2_1p5bEmbed,
+            //     WhichCandleAsset::QwenV2_1p5bEmbed,
             // ),
             ..Default::default()
         };
@@ -739,7 +738,7 @@ impl AgentSessionBuilderTrait for DocumentRAGSession<'_> {
             candle_embed_config.tokenizer_file = None;
             candle_embed_config.tokenizer_config_file = None;
             candle_embed_config.candle_asset =
-                Some(crate::candle_assets::candle_which::WhichCandleAsset::QwenV2_1p5bEmbed);
+                Some(WhichCandleAsset::QwenV2_1p5bEmbed);
         }
 
         // Add openAI_api if available
@@ -900,9 +899,9 @@ pub fn fields_in_schemas(lhs_schema: SchemaRef, rhs_schema: SchemaRef) -> Vec<St
 
 pub mod test_doc_rag_session {
     use super::*;
-    use crate::candle_chat::message_history::MessageHistoryBuilderTraitExt;
     use arrow::array::{ArrayRef, RecordBatch, StringArray};
     use parking_lot::RwLock;
+    use phymes_ai::candle_chat::message_history::MessageHistoryBuilderTraitExt;
     use phymes_core::{
         metrics::HashMap,
         session::{
