+++ phymes-agents/src/session_plans/tool_agent_session.rs
@@ -3,17 +3,15 @@ use std::sync::Arc;
 
 use super::agent_session_builder::AgentSessionBuilderTrait;
 #[cfg(feature = "openai_api")]
-use crate::openai_asset::chat_processor::OpenAIChatProcessor;
-use crate::{
-    candle_chat::{
-        chat_config::CandleChatConfig, chat_processor::CandleChatProcessor,
-        message_aggregator_processor::MessageAggregatorProcessor,
-        message_parser_processor::MessageParserProcessor,
-    },
-    candle_ops::{
-        ops_processor::CandleOpProcessor, ops_which::WhichCandleOps,
-        summary_config::CandleOpsSummaryConfig, summary_processor::OpsSummaryProcessor,
-    },
+use phymes_ai::openai_asset::chat_processor::OpenAIChatProcessor;
+use phymes_ai::{candle_assets::candle_which::WhichCandleAsset, candle_chat::{
+    chat_config::CandleChatConfig, chat_processor::CandleChatProcessor,
+    message_aggregator_processor::MessageAggregatorProcessor,
+    message_parser_processor::MessageParserProcessor,
+}};
+use phymes_etl::candle_ops::{
+    ops_processor::CandleOpProcessor, ops_which::WhichCandleOps,
+    summary_config::CandleOpsSummaryConfig, summary_processor::OpsSummaryProcessor,
 };
 use phymes_core::{
     metrics::ArrowTaskMetricsSet,
@@ -473,7 +471,7 @@ impl AgentSessionBuilderTrait for ToolAgentSession<'_> {
                 std::env::var("HOME").unwrap_or("".to_string())
             )),
             candle_asset: Some(
-                crate::candle_assets::candle_which::WhichCandleAsset::SmolLM2_135MChat,
+                WhichCandleAsset::SmolLM2_135MChat,
             ),
             ..Default::default()
         };
@@ -482,7 +480,7 @@ impl AgentSessionBuilderTrait for ToolAgentSession<'_> {
         #[cfg(feature = "hf_hub")]
         {
             candle_chat_config.candle_asset =
-                Some(crate::candle_assets::candle_which::WhichCandleAsset::QwenV2p5_3bChat);
+                Some(WhichCandleAsset::QwenV2p5_3bChat);
             candle_chat_config.openai_asset = None;
             candle_chat_config.weights_config_file = None;
             candle_chat_config.weights_file = None;
@@ -555,8 +553,8 @@ impl AgentSessionBuilderTrait for ToolAgentSession<'_> {
 
 pub mod test_tool_agent_session {
     use super::*;
-    use crate::candle_chat::message_history::MessageHistoryBuilderTraitExt;
     use parking_lot::RwLock;
+    use phymes_ai::candle_chat::message_history::MessageHistoryBuilderTraitExt;
     use phymes_core::{
         metrics::HashMap,
         session::{
