+++ phymes-ai/README.md
@@ -1,0 +1,11 @@
+# PHYMES: Parallel HYpergraph MEssaging Streams
+
+Artificial Intelligence (AI) crate
+
+<!--- ANCHOR: synopsis --->
+
+## Synopsis
+
+The PHYMES agent crate implements the functionality for Generative AI. Specifically, the crate implements functionality for native rust text generation inference and text embedding inference using the [candle](https://github.com/huggingface/candle) crates from [HuggingFace](https://huggingface.co/). All functionality can be GPU accelerated using CUDA and CuDNN.
+
+<!--- ANCHOR_END: synopsis --->
