+++ phymes-ai/benches/candle_asset.rs
@@ -1,6 +1,6 @@
 use candle_core::DType;
 use criterion::{Criterion, criterion_group, criterion_main};
-use phymes_agents::{
+use phymes_ai::{
     candle_assets::{
         candle_which::{WhichCandleAsset, load_model_asset_path, load_tokenizer},
         device::device,
