+++ phymes-ai/src/candle_assets/candle_asset.rs
