+++ phymes-ai/src/candle_chat/message_aggregator_processor.rs
