+++ phymes-ai/src/candle_chat/message_history.rs
