+++ phymes-ai/src/candle_chat/message_parser_processor.rs
