+++ phymes-ai/src/candle_chat/tool_parser.rs
