+++ phymes-ai/src/candle_embed/embed_document.rs
