+++ phymes-ai/src/candle_embed/embed_processor.rs
