+++ phymes-ai/src/candle_embed/mod.rs
