+++ phymes-ai/src/candle_models/mod.rs
@@ -1,0 +1,2 @@
+pub mod quantized_bert;
+pub mod quantized_qwen2;
