+++ phymes-ai/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod candle_assets;
+pub mod candle_chat;
+pub mod candle_embed;
+pub mod candle_models;
+pub mod openai_asset;
