+++ phymes-ai/src/openai_asset/chat_completion.rs
