+++ phymes-ai/src/openai_asset/chat_processor.rs
