+++ phymes-ai/src/openai_asset/common.rs
