+++ phymes-ai/src/openai_asset/embed_processor.rs
