+++ phymes-ai/src/openai_asset/embedding.rs
