+++ phymes-ai/src/openai_asset/types.rs
